charms/helm-orchestrator/requirements.txt
ops>=2.11
ops-lib-helm>=0.2
lightkube
lightkube-models
jinja2
pyyaml
